chore: rename eth-provider-types.ts to json-rpc-provider-types.ts
